Don't restrict org outline path length
Limiting candidate length doesn't seem to be used in other sources, and
doesn't seem useful here. The leaf node is likely the most interesting
part of the candidate.

Also helm-window seems to be not yet created when calling
helm-org--get-candidates-in-file, trying to determine the width for it
will then determine it for the current window instead, which might be
much narrower.